requirements.txt
Flask==2.3.3
requests==2.31.0
celery==5.3.4
redis==5.0.0
flower==2.0.1
python_dotenv==1.0.1